+++ Cryptography/gcd.h
@@ -9,13 +9,14 @@
 #define gcd_h
 
 #include <stdio.h>
+#include <stdint.h>
 
 void scan_for_gcd(int *a, int *b);
 
-void print_gcd(long long a, long long b);
+void print_gcd(int64_t a, int64_t b);
 
 int gcd(int a, int b);
 
-void extended_gcd(long long a, long long b, long long *gcd, long long *m, long long *n);
+void extended_gcd(int64_t a, int64_t b, int64_t *gcd, int64_t *m, int64_t *n);
 
 #endif /* gcd_h */
