+++ Cryptography/square-and-multiply.c
@@ -8,15 +8,15 @@
 #include "square-and-multiply.h"
 
 void print_for_exponent(int x, int e, int p) {
-	long long result = exponent(x, e, p);
+	int64_t result = exponent(x, e, p);
 	printf("%d^%d (mod %d) === %lld\n", x, e, p, result);
 }
 
 // x^e (mod p)
-long long exponent(long long x, long long y, long long mod) {
+int64_t exponent(int64_t x, int64_t y, int64_t mod) {
 	if (y == 0)
 		return 1;
-	long long e = exponent(x, y / 2, mod) % mod;
+	int64_t e = exponent(x, y / 2, mod) % mod;
 	e = (e * e) % mod;
 	if (y % 2 == 1)
 		e = (e * x) % mod;
